Flash message on report
