+++ src/dpr/components/_reports/report-page/view.njk
@@ -3,7 +3,7 @@
 {% from "./report-template/view.njk" import dprReportTemplate %}
 {% from "../../validation-summary/view.njk" import dprValidationSummary %}
 
-{% macro dprReportPage(renderData, validationErrors) %}
+{% macro dprReportPage(renderData, validationErrors, subscriptionMessages) %}
   {% set printable = renderData.printable %}
   <div class="govuk-width-container">
     <div id="dpr-report-page" class="dpr-report-page govuk-width-container report-list-container">
@@ -11,7 +11,7 @@
 
       <div class="dpr-report-print-message-container">{{ dprReportPrintMessage(printable) }}</div>
 
-      <div class="dpr-report-page__heading-container">{{ dprReportHeading(renderData, validationErrors) }}</div>
+      <div class="dpr-report-page__heading-container">{{ dprReportHeading(renderData, validationErrors, subscriptionMessages) }}</div>
 
       <div class="dpr-report-page__template-container">{{ dprReportTemplate(renderData) }}</div>
