+++ src/dpr/middleware/setUpDprResources.ts
@@ -31,6 +31,8 @@ export const setupResources = (
   return async (req, res, next) => {
     populateValidationErrors(req, res)
 
+    populateSubsMessages(req, res)
+
     try {
       // 1. Initialise our feature flags to app.locals
       await setFeatureFlags(res, services.featureFlagService)
@@ -98,6 +100,18 @@ const populateValidationErrors = (req: Request, res: Response) => {
   const errors = req.flash(`DPR_ERRORS`)
   if (errors && errors[0]) {
     res.locals['validationErrors'] = JSON.parse(errors[0])
+  }
+}
+
+const populateSubsMessages = (req: Request, res: Response) => {
+  const subscribedMessage = req.flash(`DPR_SUBSCRIBED`)
+  if (subscribedMessage && subscribedMessage[0]) {
+    res.locals['subscribedMessage'] = JSON.parse(subscribedMessage[0])
+  }
+
+  const unsubscribedMessage = req.flash(`DPR_UNSUBSCRIBED`)
+  if (unsubscribedMessage && unsubscribedMessage[0]) {
+    res.locals['unsubscribedMessage'] = JSON.parse(unsubscribedMessage[0])
   }
 }
