+++ src/dpr/routes/journeys/view-report/report.njk
@@ -2,5 +2,5 @@
 {% extends layoutPath %}
 
 {% block content %}
-  {{ dprReportPage(renderData, validationErrors) }}
+  {{ dprReportPage(renderData, validationErrors, subscriptionMessages) }}
 {% endblock %}
